requirements.txt
psycopg
psycopg-binary
psycopg-pool
python-dateutil
pytrading212
dataclasses_json
tenacity
pyTelegramBotAPI